+++ apollo-router/src/services/http/tests.rs
@@ -342,3 +342,95 @@ async fn test_subgraph_h2c() {
         r#"{"data":null}"#
     );
 }
+/*
+// starts a local server emulating a subgraph returning compressed response
+async fn emulate_subgraph_compressed_response(listener: TcpListener) {
+    async fn handle(request: http::Request<Body>) -> Result<http::Response<Body>, Infallible> {
+        // Check the compression of the body
+        let mut encoder = GzipEncoder::new(Vec::new());
+        encoder
+            .write_all(
+                &serde_json::to_vec(&Request::builder().query("query".to_string()).build())
+                    .unwrap(),
+            )
+            .await
+            .unwrap();
+        encoder.shutdown().await.unwrap();
+        let compressed_body = encoder.into_inner();
+        assert_eq!(
+            compressed_body,
+            hyper::body::to_bytes(request.into_body())
+                .await
+                .unwrap()
+                .to_vec()
+        );
+
+        let original_body = Response {
+            data: Some(Value::String(ByteString::from("test"))),
+            ..Response::default()
+        };
+        let mut encoder = GzipEncoder::new(Vec::new());
+        encoder
+            .write_all(&serde_json::to_vec(&original_body).unwrap())
+            .await
+            .unwrap();
+        encoder.shutdown().await.unwrap();
+        let compressed_body = encoder.into_inner();
+
+        Ok(http::Response::builder()
+            .header(CONTENT_TYPE, APPLICATION_JSON.essence_str())
+            .header(CONTENT_ENCODING, "gzip")
+            .status(StatusCode::OK)
+            .body(compressed_body.into())
+            .unwrap())
+    }
+
+    let make_svc = make_service_fn(|_conn| async { Ok::<_, Infallible>(service_fn(handle)) });
+    let server = Server::from_tcp(listener).unwrap().serve(make_svc);
+    server.await.unwrap();
+}
+
+#[tokio::test(flavor = "multi_thread")]
+async fn test_compressed_request_response_body() {
+    let listener = std::net::TcpListener::bind("127.0.0.1:0").unwrap();
+    let socket_addr = listener.local_addr().unwrap();
+    tokio::task::spawn(emulate_subgraph_compressed_response(listener));
+    let subgraph_service = SubgraphService::new(
+        "test",
+        false,
+        None,
+        Notify::default(),
+        HttpServiceFactory::from_config("test", &Configuration::default(), Http2Config::Enable),
+    )
+    .expect("can create a SubgraphService");
+
+    let url = Uri::from_str(&format!("http://{socket_addr}")).unwrap();
+    let resp = subgraph_service
+        .oneshot(SubgraphRequest {
+            supergraph_request: supergraph_request("query"),
+            subgraph_request: http::Request::builder()
+                .header(HOST, "rhost")
+                .header(CONTENT_TYPE, APPLICATION_JSON.essence_str())
+                .header(CONTENT_ENCODING, "gzip")
+                .uri(url)
+                .body(Request::builder().query("query".to_string()).build())
+                .expect("expecting valid request"),
+            operation_kind: OperationKind::Query,
+            context: Context::new(),
+            subgraph_name: String::from("test").into(),
+            subscription_stream: None,
+            connection_closed_signal: None,
+            query_hash: Default::default(),
+            authorization: Default::default(),
+        })
+        .await
+        .unwrap();
+    // Test the right decompression of the body
+    let resp_from_subgraph = Response {
+        data: Some(Value::String(ByteString::from("test"))),
+        ..Response::default()
+    };
+
+    assert_eq!(resp.response.body(), &resp_from_subgraph);
+}
+*/
